created user controller sign up and script test-watch
