+++ src/app.js
@@ -2,43 +2,14 @@ import express from "express";
 import cors from "cors";
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
-
 import connection from "./database/database.js";
+import * as userControllers from './controllers/userControllers.js'
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 
-app.post("/sign-up", async (req, res) => {
-  try {
-    const { name, email, password } = req.body;
-
-    if (!name || !email || !password) {
-      return res.sendStatus(400);
-    }
-
-    const existingUserWithGivenEmail = await connection.query(
-      `SELECT * FROM "users" WHERE "email"=$1`,
-      [email]
-    );
-
-    if (existingUserWithGivenEmail.rows[0]) {
-      return res.sendStatus(409);
-    }
-
-    const hashedPassword = bcrypt.hashSync(password, 12);
-
-    await connection.query(
-      `INSERT INTO "users" ("name", "email", "password") VALUES ($1, $2, $3)`,
-      [name, email, hashedPassword]
-    );
-
-    res.sendStatus(201);
-  } catch (err) {
-    console.error(err);
-    res.sendStatus(500);
-  }
-});
+app.post("/sign-up", userControllers.signUp);
 
 app.post("/sign-in", async (req, res) => {
   try {
